Revert "KAFKA-15774: refactor windowed stores to use StoreFactory (#14708)"
This reverts commit 39cacca89b0396398a783307e77df31b7ebf9e5e.
